NFT
Add tracking struct mapped to user addresses.  struct contains NFT Level as
Bytes32 and a bool stating if the user at that address owns any of the NFT
Levels.  Intention is for users to be only able to Mint one NFT per owner.
Will need to account for this tracking struct when doing transfers so this is
a WIP