Loans V14
This iteration is close to what we are looking for for an initial loan
product.  This version calculates interest via APY with multiple loan
durations.  Longer Durations before repayment is required come with
higher interest.  You are rewarded with less interest/fees if you pay
back early with a minimum fee if you pay back before fees can be accrued